+++ NEWS.md
@@ -1,6 +1,13 @@
 # tern 0.9.11.9000
 
 ### Enhancements
+* Added `safe_2x2_table()` to construct 2 x 2 x k contingency tables safely.
+* Added `assert_prop_data()` to validate responder, group, and optional
+  stratification data used in proportion analyses.
+
+# tern 0.9.11
+
+### Enhancements
 * Updated `g_forest()` to support point estimates and confidence intervals
   stored in a single column. (#1499)
 * Added the `exclude_rows` argument to `g_forest()` to allow excluding selected